+++ build_runner/test/common/package_graphs.dart
@@ -2,7 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// ignore: implementation_imports
 import 'package:build_runner/src/internal.dart';
 import 'package:package_config/package_config.dart';